backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pdf2image==1.16.3
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
python-magic==0.4.27
pydantic==2.4.2
langdetect==1.0.9
scikit-image==0.21.0
imutils==0.5.4
python-Levenshtein==0.21.1